requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.41.0
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-summernote==0.8.11.6
idna==2.8
jsonfield==3.1.0
mysqlclient==1.4.6
oauthlib==3.1.0
peewee==3.13.1
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
swapper==1.1.2.post1
urllib3==1.25.8
